# backend/gateway/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 Go 微服务架构的智能人才运营平台 API。Swagger 注释文档以 API Gateway 对外暴露的接口为准。\n认证接口以外的大多数后台接口需要在请求头中携带 JWT：Authorization: Bearer \u003ctoken\u003e。",
        "title": "智能人才运营平台 API",
        "contact": {},
        "version": "1.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询职位投递列表，talent_id 可传 me 表示当前登录用户。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "投递"
                ],
                "summary": "获取投递列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "职位 ID",
                        "name": "job_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "人才 ID 或 me",
                        "name": "talent_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "投递状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前登录求职者投递职位。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "投递"
                ],
                "summary": "创建职位投递",
                "parameters": [
                    {
                        "description": "投递参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerAny"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户聊天会话列表。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "获取会话列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "与指定用户创建或获取已有会话。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "创建或获取会话",
                "parameters": [
                    {
                        "description": "会话参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerAny"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/evaluations": {
            "get": {
                "description": "查询 AI 评估记录。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评估"
                ],
                "summary": "获取评估记录列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "评估类型",
                        "name": "eval_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/interviews": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询面试安排列表。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "面试"
                ],
                "summary": "获取面试列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "日期",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "面试官 ID",
                        "name": "interviewer_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "候选人 ID",
                        "name": "candidate_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建候选人面试安排。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "面试"
                ],
                "summary": "创建面试安排",
                "parameters": [
                    {
                        "description": "面试安排",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerAny"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/interviews/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取面试数量和状态统计。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "面试"
                ],
                "summary": "获取面试统计",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/jobs": {
            "get": {
                "description": "公开查询职位列表，支持关键词、状态、地点、学历、级别等筛选。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "获取职位列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "职位状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "职位类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "地点",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "学历",
                        "name": "education",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "级别",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "经验要求",
                        "name": "experience",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方向",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建招聘职位，需要管理员或 HR 类角色。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位"
                ],
                "summary": "创建职位",
                "parameters": [
                    {
                        "description": "职位信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerAny"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/logs/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取操作日志聚合统计。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志"
                ],
                "summary": "获取日志统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.SwaggerAny": {
            "type": "object",
            "additionalProperties": true
        },
...
        "main.SwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "InternalToken": {
            "type": "apiKey",
            "name": "X-Internal-Token",
            "in": "header"
        }
    }
}